Smooth Pointlights movements

+++ Engine/AtlasEngineGI/include/data/model.h
@@ -33,7 +33,7 @@ class Model
     inline GLuint getShaderTypeIndex() const{return m_material->getShaderTypeIndex();}
     Mesh *getMesh(const GLuint &index) const{return m_meshes[index];}
     Material *getMaterial() const{return m_material;}
-    inline GLuint numberOfMeshes() const{return m_meshes.size();}
+    inline size_t numberOfMeshes() const{return m_meshes.size();}
     
     //  Setters
     inline void setTransform(const glm::mat4 &transform){m_transform = transform;}
